requirements.txt
altgraph==0.17
autopep8==1.4.4
future==0.18.2
pefile==2019.4.18
Pillow==7.0.0
pycodestyle==2.5.0
PyInstaller==3.6
pyocr==0.7.2
PySimpleGUI==4.15.2
pywin32-ctypes==0.2.0
